Fixes. I think the State structure size changed in remill, so old snapshots will need to be updated

+++ runtime/klee-libc/strlen.c
@@ -13,5 +13,5 @@ size_t strlen(const char *str) {
   const char *s = str;
   while (*s)
     ++s;
-  return s - str;
+  return (size_t) (s - str);
 }
